fixing the image size integer types of the EOX_IMAGE_INFO structure to get rid of the (size_t*) to (int*) pointer casts + changing the structure identifiers to uppercase to follow the common C convetion

+++ eoxserver/processing/gdal/reftools.py
@@ -65,8 +65,8 @@ class SUBSET(C.Structure):
 
 
 class IMAGE_INFO(C.Structure):
-    _fields_ = [("x_size", C.c_size_t),
-                ("y_size", C.c_size_t),
+    _fields_ = [("x_size", C.c_int),
+                ("y_size", C.c_int),
                 ("geotransform", C.ARRAY(C.c_double, 6))]
